Refactor tests and add test for /stats Move test data into its own new file data.py in test folder Separate out Medicine tests into a new file Add test for statistics and fix bug in stats Add link to Advanced editor, separate pages tests
