+++ tests/test_templates.py
@@ -1,59 +1,74 @@
-from datetime import datetime as dt
 from django.test import TestCase, Client
-from django.contrib.auth.models import User
-from sau.models import Sheep, Farm, Medicine, Dose
-
-def generate_sheep_db():
-    s = Sheep.objects.create(name='britanna',
-                             birth_date_utc=dt.now(),
-                             sex='f',
-                             farm=Farm.objects.all()[0])
-    s.save()
+from sau.models import Sheep
+from .data import generate_test_db
 
 
 class TemplateTestcase(TestCase):
-
-    @classmethod
-    def setUpClass(cls):
-        farm = Farm.objects.create(name='Misty Mountains')
-        cls.username = 'Smeagol'
-        cls.password = 'myprecious'
-        cls.user = User.objects.create_user(username=cls.username,
-                                             password=cls.password)
-        farm.farmers.add(cls.user)
-        farm.save()
-        generate_sheep_db()
-
     def setUp(self):
-        self.c = Client()
+        generate_test_db()
 
+    def login(self, do_login=True):
+        c = Client()
+        if not do_login:
+            return c
+        login = c.login(username='Smeagol', password='myprecious')
+        self.assertTrue(login)
+        return c
+
+    def _get_html(self, client, url, **kwargs):
+        response = client.get(url, **kwargs)
+        self.assertEqual(response.status_code, 200)
+        return str(response.content)
 
     def test_nologin_site(self):
+        client = self.login(do_login=False)
         for url in ('/', '/sau/britanna', '/sau/nosuchsau'):
-           response = self.c.get(url, follow=True)
-           self.assertEqual(response.status_code, 200, msg='on url=%s'%url)
-           self.assertIn('Django admin', str(response.content))  # go to admin
+            response = client.get(url, follow=True)
+            self.assertEqual(response.status_code, 200, msg='on url=%s' % url)
+            self.assertIn('Django admin', str(response.content))  # go to admin
 
-        for url in ('nosuchpage',):
-           response = self.c.get(url, follow=True)
-           self.assertEqual(response.status_code, 404, msg='on url=%s'%url)
-
+        for url in ('nosuchpage', ):
+            response = client.get(url, follow=True)
+            self.assertEqual(response.status_code, 404, msg='on url=%s' % url)
 
     def test_login_wrong(self):
-        login = self.c.login(username='nouser', password='nopass')
-        response = self.c.get('/sau/britanna', follow=True)
+        client = self.login(do_login=False)
+        login_ = client.login(username='nouser', password='nopass')
+        self.assertFalse(login_)
+        response = client.get('/sau/britanna', follow=True)
         self.assertEqual(response.status_code, 200)
         self.assertIn('Django admin', str(response.content))  # go to admin
 
     def test_login_site(self):
-        login = self.c.login(username=self.username, password=self.password)
-        self.assertEqual(True, login)
-        response = self.c.get('/', follow=True)
+        client = self.login()
+        response = client.get('/', follow=True)
         self.assertEqual(response.status_code, 200)
         self.assertIn('Hei', str(response.content))
-        self.assertIn(self.username, str(response.content))
+        self.assertIn('Smeagol', str(response.content))
 
+    def test__self_setup(self):
+        self.assertEqual(9, len(Sheep.objects.all()))
 
-    @classmethod
-    def tearDownClass(cls):
-        pass
+    def test_index(self):
+        client = self.login(do_login=True)
+        html = self._get_html(client, '/')
+        self.assertIn('2608', html)
+        self.assertIn('lolcakes', html)
+        self.assertIn('britanna', html)
+
+    def test_sau_statistics(self):
+        client = self.login()
+        html = self._get_html(client, '/sau/britanna')
+        self.assertIn('min: 10', html)
+        self.assertIn('max: 12', html)
+        self.assertEqual(3, html.count('<svg'))
+
+    def test_stats(self):
+        client = self.login()
+        html = self._get_html(client, '/stats')
+        self.assertEqual(2, html.count('<svg'))
+
+    def test_sau_edit(self):
+        client = self.login()
+        html = self._get_html(client, '/sau/britanna/edit')
+        self.assertIn('Advanced editor', html)
